refactor(payment): paymentplatform.mock 디렉토리 제거 및 테스트 헬퍼 위치 정리
bounded context 밖 공용 mock 패키지(paymentplatform.mock)를 폐기하고
모든 Fake를 payment.mock 패키지로 통일한다.

- FakeIdempotencyStore: paymentplatform.mock → payment.mock 이동
- PaymentCheckoutServiceImplTest: import 교체 (paymentplatform.mock → payment.mock)
- FakeStockCachePortAtomicTest: payment.mock → payment.application 이동
  (mock 디렉토리엔 Fake 구현체만 두는 규칙 준수)
- paymentplatform.mock 빈 디렉토리 제거
- PLAN.md Task 5 완료 기록, STATE.md stage를 execute → ship으로 전환

./gradlew test: 512 passed, 0 failed
