+++ docs/STATE.md
@@ -5,8 +5,8 @@
 ## 활성 작업
 
 - **주제**: CLEANUP-BATCH-C (코드 레벨 정리 — 미사용 코드 제거 + 워커 보일러플레이트 헬퍼화 + 테스트 헬퍼 위치 통일)
-- **단계**: execute
-- **활성 태스크**: Task 5: paymentplatform.mock 디렉토리 위치 정리
+- **단계**: ship
+- **활성 태스크**: 전 태스크 완료 (Task 1~5)
 - **이슈/브랜치**: #98
 - **파일**: docs/topics/CLEANUP-BATCH-C.md / docs/CLEANUP-BATCH-C-PLAN.md
