+++ payment-service/src/test/java/com/hyoguoo/paymentplatform/payment/application/FakeStockCachePortAtomicTest.java
@@ -1,9 +1,10 @@
-package com.hyoguoo.paymentplatform.payment.mock;
+package com.hyoguoo.paymentplatform.payment.application;
 
 import com.hyoguoo.paymentplatform.payment.application.port.out.StockCompensationAtomicResult;
 import com.hyoguoo.paymentplatform.payment.application.port.out.StockDecrementAtomicResult;
 import com.hyoguoo.paymentplatform.payment.domain.PaymentOrder;
 import com.hyoguoo.paymentplatform.payment.domain.enums.PaymentOrderStatus;
+import com.hyoguoo.paymentplatform.payment.mock.FakeStockCachePort;
 import java.math.BigDecimal;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
